customise WebRTCSession
