+++ js/modules/webrtc/qbRTCPeerConnection.js
@@ -1,0 +1,139 @@
+// cross-browser polyfill
+var RTCPeerConnection = window.RTCPeerConnection || window.webkitRTCPeerConnection || window.mozRTCPeerConnection;
+var RTCSessionDescription = window.RTCSessionDescription || window.mozRTCSessionDescription;
+var RTCIceCandidate = window.RTCIceCandidate || window.mozRTCIceCandidate;
+
+if (RTCPeerConnection) {
+
+RTCPeerConnection.prototype.init = function(service, options) {
+  this.service = service;
+  this.sessionID = options && options.sessionID || Date.now();
+  this.type = options && options.description ? 'answer' : 'offer';
+
+  this.addStream(this.service.localStream);
+  this.onicecandidate = this.onIceCandidateCallback;
+  this.onaddstream = this.onRemoteStreamCallback;
+  this.onsignalingstatechange = this.onSignalingStateCallback;
+  this.oniceconnectionstatechange = this.onIceConnectionStateCallback;
+
+  if (this.type === 'answer') {
+    this.onRemoteSessionCallback(options.description, 'offer');
+  }
+};
+
+RTCPeerConnection.prototype.getSessionDescription = function(callback) {
+  var self = this;
+
+  if (self.type === 'offer') {
+    // Additional parameters for SDP Constraints
+    // http://www.w3.org/TR/webrtc/#constraints
+    // self.createOffer(successCallback, errorCallback, constraints)
+    self.createOffer(successCallback, errorCallback);
+  } else {
+    self.createAnswer(successCallback, errorCallback);
+  }
+
+  function successCallback(desc) {
+    self.setLocalDescription(desc, function() {
+      callback(null, desc);
+    }, errorCallback);
+  }
+  function errorCallback(error) {
+    callback(error, null);
+  }
+};
+
+RTCPeerConnection.prototype.onIceCandidateCallback = function(event) {
+  var candidate = event.candidate;
+
+  if (candidate) {
+    trace("onICECandidate: " + JSON.stringify(candidate));
+
+    this.iceCandidates = this.iceCandidates || [];
+    this.iceCandidates.push({
+      sdpMLineIndex: candidate.sdpMLineIndex,
+      sdpMid: candidate.sdpMid,
+      candidate: candidate.candidate
+    });
+  }
+};
+
+// handler of remote session description
+RTCPeerConnection.prototype.onRemoteSessionCallback = function(sessionDescription, type) {
+  var desc = new RTCSessionDescription({sdp: sessionDescription, type: type});
+  this.setRemoteDescription(desc);
+};
+
+// handler of remote media stream
+RTCPeerConnection.prototype.onRemoteStreamCallback = function(event) {
+  if (typeof this.service.onRemoteStreamListener === 'function'){
+    this.service.onRemoteStreamListener(event.stream);
+  }
+};
+
+RTCPeerConnection.prototype.addCandidates = function(iceCandidates) {
+  var candidate;
+  for (var i = 0, len = iceCandidates.length; i < len; i++) {
+    candidate = {
+      sdpMLineIndex: iceCandidates[i].sdpMLineIndex,
+      sdpMid: iceCandidates[i].sdpMid,
+      candidate: iceCandidates[i].candidate
+    };
+    this.addIceCandidate(new RTCIceCandidate(candidate));
+  }
+};
+
+RTCPeerConnection.prototype.onSignalingStateCallback = function() {
+  // send candidates
+  if (this.signalingState === 'stable' && this.type === 'offer'){
+    this.service._sendCandidate(this.opponentId, this.iceCandidates);
+  }
+};
+
+RTCPeerConnection.prototype.onIceConnectionStateCallback = function() {
+  trace("onIceConnectionStateCallback: " + this.iceConnectionState);
+
+  var newIceConnectionState = this.iceConnectionState;
+
+  // read more about all states:
+  // http://w3c.github.io/webrtc-pc/#idl-def-RTCIceConnectionState
+  //
+  // 'disconnected' happens in a case when a user has killed an application (for example, on iOS/Android via task manager).
+  // So we should notify our user about it.
+
+  // notify user about state changes
+  //
+  if(typeof this.service.onSessionConnectionStateChangedListener === 'function'){
+  	var sessionState = null;
+  	if (newIceConnectionState === 'checking'){
+        sessionState = this.service.SessionConnectionState.CONNECTING;
+  	} else if (newIceConnectionState === 'connected'){
+        sessionState = this.service.SessionConnectionState.CONNECTED;
+  	} else if (newIceConnectionState === 'failed'){
+        sessionState = this.service.SessionConnectionState.FAILED;
+  	} else if (newIceConnectionState === 'disconnected'){
+        sessionState = this.service.SessionConnectionState.DISCONNECTED;
+  	} else if (newIceConnectionState === 'closed'){
+        sessionState = this.service.SessionConnectionState.CLOSED;
+  	}
+
+  	if(sessionState != null){
+      this.service.onSessionConnectionStateChangedListener(sessionState);
+    }
+  }
+
+  //
+  if (newIceConnectionState === 'closed'){
+    //peer = null;
+  }
+};
+
+function trace(text) {
+  // if (config.debug) {
+    console.log('[QBWebRTC]:', text);
+  // }
+}
+
+}
+
+module.exports = RTCPeerConnection;
