+++ js/modules/webrtc/qbWebRTCSession.js
@@ -5,6 +5,13 @@
  *
  */
 
+
+ /**
+  * Creates a session
+  * @param {number} An ID if the call's initiator
+  * @param {array} An array with opponents
+  * @param {enum} Type of a call
+  */
 function WebRTCSession(initiatorID, opponentsIDs, callType) {
   this.ID = generateUUID();
   this.state = this.state.NEW;
@@ -14,6 +21,46 @@ function WebRTCSession(initiatorID, opponentsIDs, callType) {
   this.callType = callType;
   //
   this.peerConnections = {};
+}
+
+/**
+ * Initiate a call
+ * @param {array} A map with custom parameters
+ */
+WebRTCProxy.prototype.call = function(extension) {
+
+}
+
+/**
+ * Accept a call
+ * @param {array} A map with custom parameters
+ */
+WebRTCProxy.prototype.accept = function(extension) {
+
+}
+
+/**
+ * Reject a call
+ * @param {array} A map with custom parameters
+ */
+WebRTCProxy.prototype.reject = function(extension) {
+
+}
+
+/**
+ * Stop a call
+ * @param {array} A map with custom parameters
+ */
+WebRTCProxy.prototype.stop = function(extension) {
+
+}
+
+/**
+ * Update a call
+ * @param {array} A map with custom parameters
+ */
+WebRTCProxy.prototype.update = function(extension) {
+
 }
 
 /**
