+++ samples/webrtc/quickblox.js
@@ -2589,11 +2589,6 @@ WebRTCProxy.prototype.SessionConnectionState = {
   CLOSED: 5
 };
 
-WebRTCProxy.prototype.CallType = {
-  VIDEO: 'video',
-  AUDIO: 'accept'
-};
-
 
 
 
@@ -2710,13 +2705,6 @@ Blob.prototype.download = function() {
  *
  */
 
-
- /**
-  * Creates a session
-  * @param {number} An ID if the call's initiator
-  * @param {array} An array with opponents
-  * @param {enum} Type of a call
-  */
 function WebRTCSession(initiatorID, opponentsIDs, callType) {
   this.ID = generateUUID();
   this.state = this.state.NEW;
@@ -2726,46 +2714,6 @@ function WebRTCSession(initiatorID, opponentsIDs, callType) {
   this.callType = callType;
   //
   this.peerConnections = {};
-}
-
-/**
- * Initiate a call
- * @param {array} A map with custom parameters
- */
-WebRTCProxy.prototype.call = function(extension) {
-
-}
-
-/**
- * Accept a call
- * @param {array} A map with custom parameters
- */
-WebRTCProxy.prototype.accept = function(extension) {
-
-}
-
-/**
- * Reject a call
- * @param {array} A map with custom parameters
- */
-WebRTCProxy.prototype.reject = function(extension) {
-
-}
-
-/**
- * Stop a call
- * @param {array} A map with custom parameters
- */
-WebRTCProxy.prototype.stop = function(extension) {
-
-}
-
-/**
- * Update a call
- * @param {array} A map with custom parameters
- */
-WebRTCProxy.prototype.update = function(extension) {
-
 }
 
 /**
